requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
groq==0.5.0
tiktoken==0.6.0
